Moved ANSI and Icon objects
